+++ SimpleDBServer/src/simpledb/materialize/SortPlan.java
@@ -52,7 +52,7 @@ public class SortPlan implements Plan {
     * of materializing and sorting the records.
     * @see simpledb.query.Plan#blocksAccessed()
     */
-   public int blocksAccessed() {
+   public long blocksAccessed() {
       // does not include the one-time cost of sorting
       Plan mp = new MaterializePlan(p, tx); // not opened; just for analysis
       return mp.blocksAccessed();
@@ -63,7 +63,7 @@ public class SortPlan implements Plan {
     * which is the same as in the underlying query.
     * @see simpledb.query.Plan#recordsOutput()
     */
-   public int recordsOutput() {
+   public long recordsOutput() {
       return p.recordsOutput();
    }
    
@@ -73,7 +73,7 @@ public class SortPlan implements Plan {
     * the underlying query.
     * @see simpledb.query.Plan#distinctValues(java.lang.String)
     */
-   public int distinctValues(String fldname) {
+   public long distinctValues(String fldname) {
       return p.distinctValues(fldname);
    }
